Safari-609.1.11.1/LayoutTests/js/propertyIsEnumerable-expected.txt

This tests that propertyIsEnumerable works according to the ECMA spec.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a.propertyIsEnumerable('length') is false
PASS a.propertyIsEnumerable ('foo') is true
PASS a.propertyIsEnumerable ('non-existant') is false
PASS global.propertyIsEnumerable ('aVarDecl') is true
PASS global.propertyIsEnumerable ('aFunctionDecl') is true
PASS global.propertyIsEnumerable ('Math') is false
PASS global.propertyIsEnumerable ('NaN') is false
PASS global.propertyIsEnumerable ('undefined') is false
PASS successfullyParsed is true

TEST COMPLETE
